unify all the types and update scheduler
