+++ apis/placement/v1beta1/interface.go
@@ -47,88 +47,3 @@ type PlacementObj interface {
 	PlacementSpecGetSetter
 	PlacementStatusGetSetter
 }
-
-// A BindingSpecGetSetter contains bindingSpec
-// +kubebuilder:object:generate=false
-type BindingSpecGetSetter interface {
-	GetBindingSpec() *ResourceBindingSpec
-	SetBindingSpec(*ResourceBindingSpec)
-}
-
-// A BindingStatusGetSetter contains bindingStatus
-// +kubebuilder:object:generate=false
-type BindingStatusGetSetter interface {
-	GetBindingStatus() *ResourceBindingStatus
-	SetBindingStatus(*ResourceBindingStatus)
-}
-
-var _ BindingObj = &ClusterResourceBinding{}
-var _ BindingObj = &ResourceBinding{}
-
-// A BindingObj is for kubernetes resource binding object.
-// +kubebuilder:object:generate=false
-type BindingObj interface {
-	client.Object
-	BindingSpecGetSetter
-	BindingStatusGetSetter
-}
-
-// A PolicySnapshotSpecGetSetter contains policy snapshot spec
-// +kubebuilder:object:generate=false
-type PolicySnapshotSpecGetSetter interface {
-	GetPolicySnapshotSpec() *SchedulingPolicySnapshotSpec
-	SetPolicySnapshotSpec(*SchedulingPolicySnapshotSpec)
-}
-
-// A PolicySnapshotStatusGetSetter contains policy snapshot status
-// +kubebuilder:object:generate=false
-type PolicySnapshotStatusGetSetter interface {
-	GetPolicySnapshotStatus() *SchedulingPolicySnapshotStatus
-	SetPolicySnapshotStatus(*SchedulingPolicySnapshotStatus)
-}
-
-// A PolicySnapshotObj is for kubernetes policy snapshot object.
-// +kubebuilder:object:generate=false
-type PolicySnapshotObj interface {
-	client.Object
-	PolicySnapshotSpecGetSetter
-	PolicySnapshotStatusGetSetter
-}
-
-// A PolicySnapshotSpec contains policy snapshot spec
-// +kubebuilder:object:generate=false
-type PolicySnapshotListItemGetter interface {
-	GetPolicySnapshotObjs() []PolicySnapshotObj
-}
-
-// A PolicySnapshotList is for kubernetes policy snapshot list object.
-// +kubebuilder:object:generate=false
-type PolicySnapshotList interface {
-	client.ObjectList
-	PolicySnapshotListItemGetter
-}
-
-// A ResourceSnapshotSpecGetSettter contains resource snapshot spec
-// +kubebuilder:object:generate=false
-type ResourceSnapshotSpecGetSettter interface {
-	GetResourceSnapshotSpec() *ResourceSnapshotSpec
-	SetResourceSnapshotSpec(*ResourceSnapshotSpec)
-}
-
-// A ResourceSnapshotStatusGetSetter contains resource snapshot status
-// +kubebuilder:object:generate=false
-type ResourceSnapshotStatusGetSetter interface {
-	GetResourceSnapshotStatus() *ResourceSnapshotStatus
-	SetResourceSnapshotStatus(*ResourceSnapshotStatus)
-}
-
-var _ ResourceSnapshotObj = &ClusterResourceSnapshot{}
-var _ ResourceSnapshotObj = &ResourceSnapshot{}
-
-// A ResourceSnapshotObj is for kubernetes resource snapshot object.
-// +kubebuilder:object:generate=false
-type ResourceSnapshotObj interface {
-	client.Object
-	ResourceSnapshotSpecGetSettter
-	ResourceSnapshotStatusGetSetter
-}
